Add files via upload
This excel project takes a look at all of the billionaires in the world and their characteristics. This data allowed me to create a dashboard where visuals are used to identify the richest people in the world. These charts are able to be filtered by many different characteristics such as gender, industry, country, and if the billionaire is self-made or not.